Change conversion or service/endpoint ID to Name filters (#7541)
* Change conversion or service/endpoint ID to Name fitlers

Old:
- service.id -> service.name AND call.type[0]
- endpoint.id -> service.name AND endpoint.name

New:
- service.id -> (service.name AND service.id)
- endpoint.id -> (endpoint.name AND endpoint.id)

Description:
The previous conversion was not correct/precise in all cases, because e.g. there can be even multiple services with the name name and type, but different IDs.
Also, just using 1 type is not correct, because a service can have multiple types.
Furthermore, including the service.name in the endpoint.id conversion always lead to redundant tags in the query, because in all use-cases so far (from SmartAlert links),
whenever there is a endpoint.id fitler, there is also a service.id fitler. So one and the same service.name would appear twice in the query after conversion.

With this conversion, even though it's not the most pretty, we have at least achieved the following:
- provide meta-information (in a separate *.name tag filter) what this ID is about
- have correct results (due to the user of *.id tag filter)

* Fix transfomerHelper tests

* Resolve conversion being applied in recursion

whenever the browser is refreshed, the link is passed on, or when the link is opened from a bookmark.
This is done by only doing this conversion when there is no service.name/endpoint.name with EQUALS tag already.
As a downside, this means that if the link already contains these filters with EQUALS, then we also do not do any conversion.
For whatever reason we even try to do that...

* Fix bug reported in review: name filter re-added for non-stan user

- re-introduce state to ensure that e.g. service.name filter is not just added again once removed: to ensure the conversion
  is really only ever done once
- check for tagCatalog being available, otherwise the check for hidden-tags in the expression can lead wrong results until it is loaded

* Implement review feedback