Fix infra metric charts clipping the start (#7489)
# Why

When zooming into a metric chart, sometimes the beginning and end of the line are cut off. This is confusing because it looks like the metrics are missing. The reason they are clipped at the beginning is because the ui-client is too aggressive in clipping the old metrics when the time-range is very small.

# What

Ensure that we retain at least one data point beyond the beginning of the chart so that the chart line will be rendered off the screen.

This fixes the clipping at the front. The clipping at the back is fixed by https://github.com/instana/backend/pull/12528

# References

- [Story](https://instana.kanbanize.com/ctrl_board/36/cards/75892/details/)