MDX stories are broken (#8281)
# Why

We auto-generate story titles in our `storyLoader.js`. However,
it seems that titles are propagated in a different way in MDX files
causing these stories to fail rendering.

# What

Add a dummy-title so that the stories render again. There are still
some shortcomings/problems though, e.g., the examples do not load.
However, at least the written docs are accessible again.