Mark a stable version for the ui-client only once and only on a delivery branch (#8558)
# Why

The stable version files for the backend (or ui-client) in `/mnt/efs/data/instana-version/stable/backend-<branch-name>.stable` that are accessible from `dev-jenkins` are only relevant for delivery (releasable) branches, as the value within these files is used by the corresponding `ui-client` pipeline to determine what the latest successful `backend` version when building the resulting container images, and vice versa for the `backend` pipelines determining the last successful `ui-client` version. Therefore we should only be marking the stable versions on a delivery (releasable) branch, e.g. `develop`, `release-XYZ`, etc.

Marking the stable ui-client in a separate stage is to keep somewhat behavioral consistency with the backend pipeline and also making it an **explicit** stage instead of a side effect that happens as part of the CodeBuild job.

# What

- Mark the stable version for the ui-client only once and only on a delivery branch in a separate Jenkins stage so that it's explicit instead it being a mostly unseen side effect that happens as part of the CodeBuild job.
- Update `ci-shared-tools` to the latest available `v0.0.5`: https://github.ibm.com/instana/ci-shared-tools/releases/tag/v0.0.5

# References

- Corresponding backend change: https://github.ibm.com/instana/backend/pull/13765