Do not include migrated Custom Events when using deprecated fitler (#9868)
To only show the deprecated items that require an action. Otherwise it is pretty hard
for customers to find the ones that require an action, as we would mix in the items that
already got migrated. And this was the core idea of these filters, to have a list of items
that need to be migrated.