Fast query mode (#9888)
# Why

The "Preview" toggle in UA for calls and traces has been used only internally. It enables backend to choose whether to execute a ClickHouse query against a full data set or just a subset of calls or traces.

Before we make this feature GA, we want to test it internally. This will allow us to tune the sensitivity of the sampling level selector in the backend.

# What

- rename the "Preview" toggle to "Fast query mode"
- rename the "previewEnabled" matrix parameter to "fastQueryModeEnabled"
- enable fast query mode by default
- the fast query mode should still be available in internal mode only 

# References

- https://instana.kanbanize.com/ctrl_board/66/cards/74681/details/