Add metadata for message filtering monitoring issue (#10413)
# Why

We would like to be able to sometimes restrict what traffic is processed in the infra product area, especially in our lower environments to protect them from being overloaded or for badly behaved users. Ideally we would address this by having finer grained agent keys, but until that happens, we would love to have a simple solution. See https://github.ibm.com/instana/backend/pull/16591 for more information.

# What

When messages are filtered, a monitoring issue is created associated with the host that is being filtered. This PR adds a UI for that monitoring issue.

# References

- [Story](https://instana.kanbanize.com/ctrl_board/36/cards/107286/details/)
- [Backend PR](https://github.ibm.com/instana/backend/pull/16591)