Ensure EE KPIs are consistent with series (#10572)
# Why

In EE, the KPIs sometimes don't seem to match up with the spark charts.

<img width="1142" alt="ScreenShot2022-11-18at8 37 04AM" src="https://media.github.ibm.com/user/365791/files/3365d53e-3a1e-431d-aebc-d99ff1b987cd">

e.g. the appdata-reader blocked threads shows 7.00 for one of the datapoints, but the KPI shows max 2.33. This doesn't seem right.

# What

- Use Beeinstana for calculating KPIs (rather than calculating them on the UI) since this is now fully supported.
- Fix count metrics being broken on the infra list.

# Notes

Through doing this I revealed a bug in how the count metrics are being calculated which means they no longer work fully, but this needs to be addressed in the backend.

# References

- [Story](https://instana.kanbanize.com/ctrl_board/36/cards/109446/details/)