Fixing entity of geo tags (#10635)
# Why

Tracing "Location"  `geo.*` tags are special infrastructure tags, that can be applied only to a source of a call, i.e., entity type `SOURCE`. 

The ui-client doesn't support this use-case so that if a new filter based on some location tag is created, its entity type is automatically set to `DESTINATION`, which results in invalid backend queries. If one of the location tags is used for grouping, its entity type is not set at all and the entity type dropdown is not even shown.

# What

- Only supported entity types will be considered when configuring default entity value for tag filters and grouping tags 
- The entity dropdown will show only supported entity types

# References

- https://instana.kanbanize.com/ctrl_board/66/cards/106806/details/