[containers] Update Instana base image version to v0.12.0 (#10735)
# Why & What

This is created automatically by the upstream pipeline at
https://ci.instana.io/teams/dev/pipelines/base-images:main
in order to ensure all downstream usages of the Instana
base/runtime images are kept up-to-date automatically when
the upstream base version changes.

## Important Note

This PR automatically targets the main working branch of this repository
but you can also decide to merge it to a different branch. If there are
large changes included in this version of the base/runtime images
that you are aware of, please take note to test that they are working as
expected once they are deployed in an environment especially if you have
decided to merge this PR to a release branch.