update ci-shared-tools version to upload maven artifacts to delivery.instana.io (#10842)
# Why

The `ui-client` build should now be publishing all artifacts to the new **delivery.instana.io** artifactory. The `buildspec.yml` file calls the script:

```
./build/ci-shared-tools/scripts/uploadReleaseArtifact.bash
```

but the version of this script from **ci-shared-tools** is configured to upload the Maven artifacts to the old **artifact-rnd.instana.io** artifactory.

# What

The newly created release of  **ci-shared-tools**, `v0.1.2`, changes the repository path in `uploadReleaseArtifact.bash` and should be used instead by the `ui-client` build.