show troubleshooting info box only in internal mode (#10890)
# Why
We may ask self-hosted customer to enable the troubleshooting setting and go to the Service Troubleshooting tab. Some of the charts in this tab contain explanation that links to Instana backend code. This information should not be shown to the customers.

# What
Show the explanation only in internal mode for Instana engineers and support to do troubleshooting.