update publish path for develop branch builds (#10950)
# Why
The **self-hosted** `main` branch builds are failing because it cannot find the **backend** images published by the `develop` branch builds, which are published in a different location from the release build images.  

The **ui-client** build retags the **backend** images in artifactory.  If the path to the `develop` **backend** images is updated, then the **ui-client** build will need to be update so it can find these images during the retagging step.


# What
The path to the `develop` branch images used during the retagging step will need to be updated from:

`int-docker-backend-local/backend/dev/develop` 

to 

`int-docker-backend-local/backend`

**NOTE:**
The `instanactl` tool will need to be updated by the SRE Team when **backend** PR https://github.ibm.com/instana/backend/pull/17271 is merged so that it finds the `develop` build images under the correct repo path during the deployment stage.