Hide confusing dropwizard metrics table (#11135)
# Why

All metrics associated with a JVM are shown at the bottom of the JVM dashboard as "Dropwizard Custom Metrics". This is misleading since these are not dropwizard custom metrics.

# What

Hide these metrics, which shouldn't have been shown in the first place. This was introduced here https://github.ibm.com/instana/ui-client/pull/10417.