Fix incorrect rendering in bar charts (#11347)
when non-default granularity is used, such as 15min, 20min or 120min.

This is done by allowing to explicitly set the granularity to the desired granularity,
instead of allways falling back to the next bigger default-granularities. This does not
make sense, as the data the bar-chart will render in fact is of the previously defined
granularity. So why should the chart-renderer assume something different?