Create lazy loading node with 'Try again' if loading of related calls fails (#11549)
# Why

If loading of related calls (siblings or children) fails, the lazy node disappears and it seems that there are no related calls. There is no possibility to retry.

# What

Remove the check which prevents creation of the "Try again" lazy node. This check was added temporarily, before proper error handling was implemented.

# References

- https://instana.kanbanize.com/ctrl_board/66/cards/121152/details/