Work-around to fix missing 'Load more calls' button in the lazy load call tree (#11653)
# Why

The "Load more calls" button is sometimes missing in the lazy loaded call tree, e.g. https://instana.io/s/agx76-JiS8SywzEsKgWzuA. The root cause of this issue is that ClickHouse sometimes returns incorrect value of `rows_before_limit_at_least`. 


# What

As a temporary work-around for the backend issue, we will try to retrieve one extra item, which we will then discard, but if the extra item will be returned, we will know that more items can be loaded.


# References

- https://instana.kanbanize.com/ctrl_board/66/cards/122152/details/