Update to add heliconia env and to doc how to access it (#11827)
Why
The pink environment is being phased out. A replacement environment is starting to be tested. This PR is adding the option and documentation for connecting the local ui to the heliconia backend.

What
The heliconia environment has been added as a new TARGET and an option for the backend when running yarn run dev. The development docs for installation, local development guidelines, and troubleshooting have been updated to describe connecting the new environment.