fix(settings/teams): Pagination broken, whilst removing last item of a page (#11842)
# Why
On user details, if you remove the last group on the last page, the (now) last page won't be displayed instead an empty page is presented. If there is only one page and you are on the 2nd page, then there is also no pagination.

# What
Adjust current page, when deleting the last item from a page.