fix(settings-users): empty page after last user (#11852)
# Why
After removing the last user of a page from the users list of a tenant, the user-list does not show the page before that.

# What
Doing a strategic solution - changing the page whilst removing the user.  
A refactoring of the ApiList component would breaking as the removal is handled by the parent (ApiList) and the pagination is pot. done by a renderer.