Do not add calls representing log events to search index (#12017)
# Why

When the feature flag `limitVisibleNestingLevelsInTraceEnabled` was enabled, traces with log events might not be displayed properly. 

# What

Do not add calls which represent log events to the search index. This is necessary, because calls representing a log event have the same ID as the call that they are attached to and thus the log event call would override the actual call in the index.

# References

- https://instana.kanbanize.com/ctrl_board/66/cards/127104/details/