fix(service-levels): start stairway renderer at first data point (#12108)
The stairway renderer for slo charts was attempting to shift its metric
rendering to the left for the first bucket. This resulted in an
additional datapoint at y=0, which does not work with the new
"remainingErrorBudget" metric, which does not start at 0.

This logic was a leftover from a time when we did not align the slo
chart with the actual timeConfig used in the backend. Since we do so now
there is no need to extend the first bucket into the past. This removes
that logic and with it fixes the bug for the new chart usage