fix(BigNumberKpiCard): make timeConfig adjustable (#12133)
The BigNumberKpiCard can take a full unified metric configuration to
loads its single value metric. This normally allows to adjust the
timeConfig for this metric as part of the configuration.
BigNumberKpiCard however used to always overwrite this from the
timeConfig context. This now enables the user to overwrite the
timeConfig again if so desired