Adjust time skew in eagerly loaded call trees (#12294)
# Why

If clocks on different hosts are not perfectly in sync (time skew), order of calls in a call tree might be incorrect and thus confusing. This is especially obvious if a child call starts before its parent.

Even though we already adjust the time skew for most of the spans as soon as we receive them, the issue can still occur.

# What

If a child call starts before its parent, the child call's start timestamp will be adjusted based on its parent call's start timestamp and half of the network time (request latency). The adjustment will be done only if the difference between the start timestamps is below 1,000 milliseconds, to avoid hiding of some other issues.

# References

- https://instana.kanbanize.com/ctrl_board/66/cards/127342/details/