Retry loading trace details (#12577)
# Why

Our internal [instana-devtools](https://github.ibm.com/instana/instana-devtools) helps users during development process by showing a list of all locally captured beacons and traces. By clicking on a beacon or on a trace icon, the user will be brought to Instana page load view or trace detail view, respectively. However, if the user is too fast, it may happen that loading of the beacon or trace fails due to delay in our processing pipeline. It usually takes about 10-30 seconds for a beacon or a trace to be processed and stored in our backend, before it can be displayed to the user. 

# What

- When viewing recent beacons or traces, retry the backend request if loading failed.
- Do at most 3 retries.
- Delay between retries is 15 seconds for traces and 10 seconds for beacons, because the ingestion delay for traces is currently higher.
- Retry only if the respective beacon timestamp is not older than 1 minute.
- Retry loading of trace detail view, only if the "to" timestamp of the selected time range is not older than 1 minute. 

# How to test
- install [instana-devtools](https://github.ibm.com/instana/instana-devtools)
- run ui-client locally against https://blue-instanaops.instana.io/ (all instanaops instances report beacons to blue)
- use `instana-devtools` extension when browsing https://green-instanaops.instana.io/ and jump to latest beacon/trace