Fix hanging metric selector (#12887)
# Why

The metric selector in Custom Dashboards in the Infrastructure & Platforms datasource crashes the browser and is very slow. This is due to unnecessary use of fuzzy search. Actually the results coming from the backend _should_ already be filtered by the specified query, however this doesn't appear to be the case for some reason. As a workaround, we can selectively disable fuzzy search for the metric selector.

# What

- Disable fuzzy sort in the Infrastructure metric selector overlay by adding an optional `strict` parameter to the `SelectorOverlay` component.

# References

- [Story](https://instana.kanbanize.com/ctrl_board/36/cards/135436/details/)