Avoid changing dependencies on every render (#13041)
# Why

Recent changes introduced with https://github.ibm.com/instana/ui-client/pull/12253 broke grouped logs view.

The `getColor` prop is used as a dependency of the `useEffect` hook in the `GroupedView` component. Since this prop changes with every render cycle in the `GroupedLogs` component, it leads to infinite re-render cycle between the `GroupedView` and `GroupedLogs` components. 

# What

Avoid changing the `getColor` prop with every render cycle in the `GroupedLogs` component.

# References

- https://instana.kanbanize.com/ctrl_board/66/cards/137037/details/
- https://instana.slack.com/archives/GA8SHDH99/p1692969553856599