Add conditions ready metric (#13388)
# Why

OMS would like to alert on pod ready status (e.g. if a pod is not ready for more than x minutes, send an alert). This is not currently possible with our built in alerts because the built in alerts fire with a built in duration, but OMS needs to configure the unready duration before the alert fires on a workload basis.

# What

- Add the pod condition ready metric to the ui metric catalog

# Checklist

- [x] Pull request tagged with one of the `type:*` labels
- [x] Changes are visually consistent with current components
- [ ] Add before/after screenshots
- [x] Affected areas were tested manually and everything works as expected
- [x] [Pull Request Guidelines](https://www.notion.so/instana/Pull-Request-reviews-8fd1ae2f48b84e53b1ef00accb0e52db?pvs=4) applied

# References

- [Story](https://instana.kanbanize.com/ctrl_board/36/cards/140679/details/)

# Screenshots

## After

![pod-condition-ready-alert](https://media.github.ibm.com/user/365791/files/2813a92c-52d3-4e8e-9f22-b41976d8671f)

## Before

<img width="1171" alt="Screenshot 2023-09-27 at 2 54 24 PM" src="https://media.github.ibm.com/user/365791/files/fc4516db-ced7-4ba7-a38d-01fade8e1781">