fix(service-levels): dynmically adjust indicator threshold validation (#13449)
* fix(service-levels): dynmically adjust indicator threshold validation

The threshold field is not required for all indicator variants. This
causes validation issues, if a invalid value is entered and then the
indicator is switch to a type or blueprint that does not require the
threshold field. The same applies to cases where the threshold field
would need different validation for different value types.
This updates the form logic to dynamically pick the right set of
validators for the threshold field.

The commit also includes minor clean up on default values and typing

* chore(slo-blueprint): add constant for default blueprint type

Co-authored-by: André Leifeld <andre.leifeld@ibm.com>