Always use `latest` base image (#13815)
We need to make sure CVEs are fixed in a timely manner. In order to
facilitate that, base images are now automatically rebuilt on a weekly
basis. Each `main` branch build is tagged with a semver and `latest`.
Using `latest` base images ensures each image build is always up-to-date
in terms of vulnerabilities. In the unlikely case that `latest` breaks
something, we can always temporarily switch to a previous version tagged
with a semver until the issue is resolved.