Fix merging of AP tag filter expression with RBAC's contribution filter (#13997)
# Why

New AP wizard crashes for some more complex contribution filters, e.g. filters with OR operator. The reason for that is that the `contributionFilter` is not stored in form model, but in backend model format.

# What

When merging the AP tag filter expression with the contribution filter, handle the contribution filter as a backend model.

# References

- https://instana.kanbanize.com/ctrl_board/66/cards/147453/details/