Fix table widget search (#14043)
# Why

A customer (OMS) reports that searching in the table widget returns unexpected results. This is because the form model of the query is not being correctly "enclosed" e.g. we are missing brackets around the query before joining the expression, which means that the precedence of the parts of the expression are not being constructed correctly.

# What

- Switch to using the backend form model for building the table widget query
- Also fix the bug in the `joinExpressions` logic.

# References

- [Story](https://instana.kanbanize.com/ctrl_board/36/cards/147896/details/)