Fix erroneous alerts in Infra Smart Alert Preview.
Infra smart alert preview was showing a clear mismatch between
'when alerts would have been triggered' vs 'the actual threshold violations'.

https://instana.slack.com/archives/C02FF10TDTP/p1711555452147699

It turns out that we had forgotten to include the entity type in the alert
preview API call which caused all entity types within a given metric to be
grouped together in the threshold violation calculation. This fixes that issue.