Fix infra TagFilterExpression validation (#18390)
# Why

When editing an Infrastructure custom dashboard widget, the widget becomes invalid even without modifying it. This is because we validate the filter but we do not specify that the validation should be done ignoring the entity selection (which we need to do in the infra area because the entity selection is not valid in that context). This is related to #18123.

# What

Ensure that validation is done ignoring the entity selection.

# How to test

1. Open a custom dashboard
2. Edit an infrastructure widget
3. Click confirm - the confirm button will be greyed out and the widget will not be editable

# References

- INSTA-18123
- [Release notes PR](https://github.ibm.com/instana/docs/pull/11470)