Add otel-batch-synthetic span plugin (#19709)
# Why

If a trace does not start with an entry span, a fake root span of type "batch-synthetic" will be created in the backend. If the original root spans was an OpenTelemetry span, a special handling is required and therefore another fake root span of type "otel-batch-synthetic" will be created in the backend. To display this new span type, a dedicated span plugin is needed. 

# What

- Implement a span plugin of type "otel-batch-synthetic"
- Display "Resource" attributes the same way as they are displayed for normal OpenTelemetry spans. 

# How to test

Open call details of a call with span type "otel-batch-synthetic".